Symbol table handling frozen. There are two alternatives how to handle identifiers: to have automatic variables and only one scope per function or to have declared variables and local scopes. Since I think declarations are not practical for interpreted language, I decided to go with the first alternative. That is why all symbol tables feature variables that preserve their validity throughout the function/macro.
There are two symbol tables that are generated on start-up: a global symbol
table that contains all constants (CONST_* identifiers) and all macro and
function names (initially empty), and the main symbol table that contains
all variables that are declared within the main script structure (also
initially empty). As new macros and functions are defined, new symbol
tables are appended to a linked list. The lookup function first searches
for a symbol in the symbol table of the current macro/function and then in
the global symbol table. This enables macros/functions to call other
macros/functions and to have all constants globally available. For example,
the code:

> macro func1 () {
>	print CONST_PI
> }
>
> macro func2 () {
> 	func1 ()
> }
>
> func2 ()

is perfectly ok.