This ChangeLog part consists of two pretty large changes which are completely different; the CVS server is down over the weekend, unfortunately, so I couldn't have committed each change separately. Here is the description:
When defining arrays explicitly, e.g. set a = {1,2,3}, the elements of the
array itself were evaluated during the parse. This posed no problems for
arrays in the main script flow, but it caused syntax errors in macro
definitions where elements of the defined array were variables. Consider the
following example:

  macro test () {
    set i = 1
    set a = {i, i+1, i+2}
    print a
    }

This is a perfectly legal macro definition, yet because of read-and-parse
implementation the parsing of {i, i+1, i+2} caused a syntax error because
the set directive wasn't yet evaluated and the variable i wasn't yet
assigned. This is now fixed by changing the evaluation to occur with the
statement itself and not separately.

The second change is the total implementation of the prompt () command.
This command provides input from the user.

Synopsis:

  set answer = prompt ("Question? " [, "expected_type"])

First argument is a string that is used as the prompt to which the user
has to give an answer. The answer is then parsed and returned as an
expression. Typically one whould use this command as:

  set answer = prompt ("Is this correct [Y/n]? ")

The second argument is optional and may be used to test the answer type.
The expected_type is a string that must be one of the following:

  integer, boolean, double, string

The command is completely fool-proof and will loop until the answer to the
prompt is acceptable by the lexer. The command will interrupt on empty input
or on CTRL+D.