Transformation of the flux sigma to magnitude sigma slightly improved in cases of small SNR. The functions that do the transformation have been heavily commented.
Another important fixup concerns the magnitude sigma to flux weight
transformation; the original w = 1/\sigma^2 scheme was troublesome for
small values of \sigma, because the weights were overflowing the 4-char
space allowed by WD. The weights are now rescaled to the 0.01-10.0
interval. Higher values are intentionally left empty, we use 99.0 for
\sigma = 0 values.

A smaller fixup of the same function concerns cases when \sigma's that
were passed were negative. This happened in several occasions when users
used Rucinski's 3rd RV column as standard deviations. This is of course
wrong to do, but PHOEBE now reports this problem and aborts.

Finally, the read_in_observational_data () function was phoebe_debug()'ed.