Additional small fixups and enhancements regarding sigma-to-weight trans- formation done. The function has been pretty fairly tested and should be ok for scientific use now.
There is a caveat that should be added to the docs: if one wants all
photometric and/or RV data points to be taken with equal weights, one
must not use standard deviations in the magnitude regime, since the
transformation depends on the individual flux value. Thus weights must
be used and set to an equal value, e.g. 1.0.