I just learned that errno value is not preserved after any IO call, including printf; from the documentation:
A common mistake is to do

      if (somecall() == -1) {
          printf("somecall() failed\n");
          if (errno == ...) { ... }
      }

where  errno  no longer needs to have the value it had upon return from
somecall() (i.e., it may have been changed by the  printf()).   If  the
value  of  errno  should be preserved across a library call, it must be
saved:

      if (somecall() == -1) {
          int errsv = errno;
          printf("somecall() failed\n");
          if (errsv == ...) { ... }
      }

This is now fixed in phoebe_open_directory and phoebe_close_directory
functions.