Another significant change has been incorporated. This time, handling the results of minimization has been rewritten. The scripter no longer returns just a chi2 value of the fit, but the whole read-only structure. For example:
	set result = minimize_using_dc ()
	print result

will print out all fields of this new structure. It is a generic structure
that is independent of the minimizer used, e.g. all minimizers return the
same structure. This minimizer feedback structure contains the following
fields:

	algorithm .. Name of the minimization algorithm
	iters     .. Number of performed iterations
	cputime   .. Computational time cost in seconds
	pars      .. A list of adjusted parameters
	initvals  .. A list of initial parameter values
	newvals   .. A list of converged parameter values
	ferrors   .. A list of formal parameter errors
	chi2s     .. A list of passband chi2 values
	wchi2s    .. A list of weighted passband chi2 values
	cfval     .. Cost function value (master chi2)

Structure fields are accessed by using the '.' operator:

	print result.algorithm
	print result.newvals
	... [etc] ...

Since the minimizer feedback now contains a wealth of information,
automatic updating of corrected values has finally been abandoned. To
update the values, use the function adopt_minimizer_results (); see
below.

Full documentation for this new principle has been written and added to
the Scripter API; please refer to that document for details.

Furthermore, two new scripter commands have been added:

	set var = compute_chi2 ([curve_index])

and

	adopt_minimizer_results (result)

The first command computes weighted chi2 values of the whole set and
returns an array of these values. If an optional argument curve_index
is passed, then chi2 of only that passband is computed.

The second command adopts the new (corrected) results of the minimizer
feedback structure 'result'. This way the user may test the results for
convergence and decide if he wants to keep or discard the new values.

This roughly sums up a ton of work. ;) The back-end had to change, since
now structures need to be supported too. Both NMS and DC minimizers were
significantly enhanced, but there are still several glitches that need
to be worked out - #warning-s hav been added to critical places. The
API documentation has been updated to properly explain all these changes.

The new implementation has been tested for the last 3 days to a reasonable
extent. It should be safe to use it for regular work, but some caution
should still be exercised and, as usual, any encountered problems reported
as soon as possible. O:)