Interstellar extinction treatment in PHOEBE changed; in the old version (0.2x) observational data were de-reddened according to Allen's formula and the model wasn't changed in any way. Strictly speaking, this constitutes data temporing and as such is not desireable. To remedy this issue, a new parameter is introduced:
	phoebe_extinction

This is a passband-dependent parameter that applies the extinction to
the computed flux, thus reddening synthetic data. There is no assumption
on the reddening law of any kind, passband values are decoupled between
themselves. This allows for user-specified extinction laws.

The units of phoebe_extinction parameter are magnitudes for the particular
passband:

	m_observed^i = m_intrinsic^i + A^i	for the i-th passband

Or, in flux units:

	j_observed^i = j_intrinsic^i *  10^(-0.4*A^i)
                     = j_intrinsic^i * exp (-0.4*ln(10)*A^i)

This correction applies to the whole passband uniquely, so it is a
simplified treatment of the rigorously developed formalism for the
2004 paper. Once the spectra will have entered the modeling, this too
will be accomodated for.

An example of using Allen's formula to preserve agreement with the old
PHOEBE version:

	A^i = 0.68 * R * E(B-V) * (1000/lambda[nm] - 0.35)

For the i-th passband with its effective wavelength lambda, given in
nanometers, these values of phoebe_extinction[i] will match the data
that has *not* been subjected to de-reddening, yet the flux levels will
be different.

Formally, extinction plays the same role as passband luminosities and
there is a 100% correlation between them. However, this is still very
useful, since if there is an extinction law that couples extinction
coefficients between different passbands, passband luminosities will
then correspond to *intrinsic* (rather than reddened) luminosities.