A very, VERY important function was added to phoebe_calculations: phoebe_interpolate (). This is a general N-dimensional linear interpolator that does the job properly. It has been optimized for speed and generality. It currently knows how to interpolate scalar values and spectra, and it is trivial to extend it to any other type supported by PHOEBE.
Another important change is that querying spectra repository now
ignores all spectra with non-zero velocities. Why? Because it just
doesn't make sense to have different velocities in the repository
since PHOEBE can compute rotational broadening on the fly.

Finally, querying spectra from the repository used a poor interpolation
algorithm that has now been replaced with the new spiffy one. Everything
checks out. The program stopped before if it encountered a non-existing
spectrum in the repository, whereas now it aborts and notifies the user
of the problem.

The job is not quite done yet: spectra are still identified by their
filenames, which is not good. This should be changed before the release
of 0.30.