Added dialogs for opening and saving Phoebe's parameter files. I placed the gui_open(save)_parameter_file() functions in phoebe_gui_accessories.c.
I create the dialogs at runtime (not from glade files) because it's more 
convenient. The "save" option works like "save as" at the moment, 
but this can easily be fixed later.

Playing with the parameter files, I noticed that 
gui_set_values_to_widgets() needs more work: if the LC and RV treeviews 
are ampty, they need to be populated with the stuff from the parameter 
file... Working on it already.