Several changes, most combobox related:
1. Removed radio buttons and placed combos instead. Added a combo for
phoebe_passband_mode, although I don't know what it's for.

2. Checked and double-checked that all the combos have the same options
as their parameters. The string that appears in the combo must match
exactly its parameter option, or the gui_set_value_to_widget() function
will ignore it, or set it the wrong way. As a consequence, the GUI
is slightly wider, because of long options of phoebe_model parameter.

3. Added "init" functions for lc_obsmenu and rv_obsmenu combos in the
plotting tabs. Also added "update" functions that go through the
treeviews with data and fill those combos accordingly. Something similar
could be done with zoom and offset combos, but I don't think it's
essential at the moment.

Finaly, I think that, with the exception of all-related-to-spots, the
parameter-widget interoperability is now complete.