Configuration handling changed. Instead of having a bunch of global variables laying around, there is now a proper way (TM) implemented. In order to add a new configuration entry, use:
	phoebe_config_entry_add (type, keyword, defval);

where:

	type is PHOEBE_type (TYPE_INT, ..., TYPE_STRING)

	keyword is a string (i.e. "PHOEBE_BASE_DIR")

	defval is the default value (i.e. "/usr/local/share/phoebe")

To set or get the value of the configuration entry, use:

	phoebe_config_entry_set (keyword,  value);

	phoebe_config_entry_get (keyword, &value);

Please take a look at phoebe_configuration.c to see the complete 
implementation. All necessary changes in the scripter and the GUI have 
been incorporated as well. The options get loaded and saved 
automatically by using:

	phoebe_config_load (char *filename);

	phoebe_config_save (char *filename);

The only thing that still remains to be done is legacy support.