Backend for accessing the repository of stellar spectra completely reworked. When I say completly, I really mean it ;) - PHOEBE now supports Munari et al. (2005) spectra. Function query_spectra_repository is obsolete and replaced by phoebe_spectra_set_repository (). This function accesses the directory given as argument and recursively builds a table of all spectra that is stored in a global PHOEBE_spectra_repository variable. I think the logic of creating a grid table is nifty, so I briefly explain it here:
1) The directory is read recursively and all filenames are parsed. The 
ones that enter the repository (no rotational velocity and no alpha/Fe) 
are added to the array of spectra in order of appearance. We omit 
rotational velocities because there is a dedicated function in PHOEBE to 
do just that. Alpha enhancement may enter some time later, but I'm not 
sure yet how realistic it is.

2) During the parsing stage, all unique T/log(g)/[M/H] are stored, each 
in a separate array within the PHOEBE_spectra_repository structure 
(these arrays are called Teffnodes, loggnodes, and metnodes).

3) Once the directory tree is traversed, we build a 3D grid based on the 
nodes obtained in step 2. All elements of this grid will be pointers to 
spectra found and stored in step 1.

4) We do a second pass over all found spectra and, knowing the geometry 
of the table, we link each table element to the corresponding spectrum.

Once all that is done, there is PHOEBE_spectra_repository.table field 
that holds the pointers to all spectra in tabular form. The file of the 
spectrum, when needed, is accessed by getting the indices i, j, k, and 
opening PHOEBE_spectra_repository.table[i][j][k]->filename. That's all 
there's to it, but the change is substantial. Multi-D interpolation has 
not been updated, that's next.