Two new parameters added: phoebe_lc_id and phoebe_rv_id. Both are string arrays and are meant to hold the common name of the light/RV curve that is used for identification. Instead of passband name the curves will be identified by their IDs. These values should be user-defineable; to demonstrate this, I added an entry to phoebe_load_lc glade file, but this entry is not connected with anything at the moment.
Next to this I think it would be useful to have a toggle button in the 
configuration window that would enforce the usage of passbands as 
identifiers (to offer the old way of referring to curves); in that case 
the entry would be grayed out and the passband name automatically copied 
once it's set. Or we could forget about this and have the users set the 
name to whatever they like.

GUI masters, please hook up a widget to phoebe_lc_id and phoebe_rv_id 
and change the tree models in combo boxes to refer to IDs rather than 
passband names.

This closes bug #1812435.