Two modifications were done to configure:
1) ldd was replaced with ${LDD:-ldd} so that a user can specify LDD= at 
   compile time, and
2) test x$plotters = x changed to test x"$plotters" = x.

Unfortunately, there is still an outstanding problem with the gnuplot 
ldd test: on linux, gnuplot is linked with /both/ libgd and libpng, but 
only picsize works, so inverting the libgd test doesn't solve the 
problem. Thus Patrick's suggestion to create a test png file and see if 
it worked would probably be best.