Scripter API documentation added to the main tree. The documentation translated from latex to docbook, which is much more convenient for technical documentation; see:
https://lists.ubuntu.com/archives/edubuntu-devel/2006-June/001553.html

for the rationale. In addition to that, drupal (phoebe homepage engine) 
knows how to convert docbook to drupal book, which will make uploading a 
whole lot easier. The first three sections have been docbook-ed. A good 
docbook manual is here, in case anyone wants to jump in and help:

http://www.docbook.org/tdg5/en/html/docbook.html

The build system has been made docbook-aware (a custom m4 macro script 
is added that is used in configure.ac). I took the base from:

http://autoconf-archive.cryp.to/ac_check_docbook_xslt.html

and adapted it slightly to our needs. The call to xml->html translator 
xsltproc is still absolute, I will fix this later. I would strongly 
recommend doing the same for the GUI documentation once we get to it.