Three important changes have been done with this commit:
1) Interstellar extinction is now passed correctly to DC;
2) Third light minimization in % properly corrects passband 
   luminosities when adjusted, and
3) most importantly, a feature long overdue: PHOEBE can now check for 
parameter sanity before the minimizer is called. In particular, the 
following functions are implemented:

  bool phoebe_parameter_is_within_limits (PHOEBE_parameter *par);

  int phoebe_parameters_check_bounds (char **offender);

The first function returns TRUE/FALSE depending on the parameter value 
(TRUE if within bounds, FALSE otherwise), and the second function sweeps 
across all adjustable parameters and, if any if found that falls out of 
bounds, non-SUCCESS is returned and offender is set to that qualifier. 
The latter is done for facilitated error handling in the drivers.