Clicking on the plot button in the LC window has been hijacked from gnuplot and is now handled by the cairo backend. So far three widgets are fully functional: the choice of the observed passband, and lower and upper x (phase or time) limits.
The implementation logic is the following. To assure consistency across 
all plot screens, a single handler gui_plot_area_init() exists that 
initializes all plot properties. It takes two arguments, the plotting 
area and the plot button. The latter is used for two purposes: 1) to 
associate all plot property widgets and 2) to initiate a plot. The 
property widgets are read out from the button data table in 
gui_plot_area_init(), and to each a signal is connected that catches the 
"changed" signal and updates the given property.

For more details see phoebe_gui_plotting, line 130, and 
phoebe_gui_types, line 116.

Cairo plotting will be AWESOME! :)