1) Merged revisions 1187 to 1130 from trunk.
2) A very simple and non-fool-proof system for running the fitting 
operations on a separate thread is implemented. The effect is that the 
main window should be fully responsive while the fitting is in progress. 

Non-fool-proof means that surely there are many ways to mess things up 
by, say, clicking the "calculate" button again while one fit is in 
progress, or by quitting the program while the fitting thread is 
running. All in my TODO list, of course!

The thing to test for is: whether the inclusion of <pthread.h> in 
phoebe_gui_callbacks.c causes problems on Windows and machines other 
than my own.