The treeview controls are now fully functional: both observations and model light curves can be plotted simultaneously. It looks pretty neat! :)
I had to comment out Patrick's tick drawing part of the code because of 
the substantially changed backend. Patrick, could you please see if 
updating to the new backend is easy and enable them back?

A word of caution: treeviews do *not* emit "row-changed" signal on 
removed rows (go figure), that is why removing LCs does not remove the 
data from the GUI_plot_data structure. I'm working on it.