A progress bar added to the bottom of the main window. It is hidden until fitting is started, and it hides again after fitting is finished. Text in the status bar gets clipped in some situations due to the positioning... I don't know how to force the progress bar to some predefined size. The speed of the progress bar can be adjusted.
There's some funny business when a minimizer exits with errors, so the 
errors don't get reported properly in the status bar and fitting tab... 
I'm on to it.

Also, it seems plotting isn't possible while fitting is in progress. I 
didn't disable it for the moment, so you can try it out, see the error. 
Maybe there's a way to solve this? Other widgets that are unsafe to use 
while fitting is in progress get disabled and enabled again after 
fitting is finished.