Polyfit command fully implemented. It takes one mandatory and one optional argument:
	set p = polyfit (curve [, indeps])

where curve is input curve to be polyfit, and indeps is an optional 
vector of phase points in which the interpolating function should be 
computed. If indeps is absent, phases are taken from curve phases.

The easiest way to apply polyfit to observed data is to:

	set c = curve ("lc_in_phases.data")
	set p = polyfit (c)

The return value of polyfit is a newly defined curve.