Important changes! Please read carefully!
(1) further support for new atmosphere tables. The changes are not propagated
    yet, but if you are interested drop me an e-mail.
(2) With the new atmosphere tables, beaming_alg can be set to full, local, global or none
    With the old tables you can only use full or none
(3) Change of defaults! Irradiator and beaming are now always true, but beaming_alg
    is now by default 'none'. This means that you can more easily switch on/off
    reflection/heating and beaming. If you want to disable beaming/reflection
    in one or more targets, while enabling it in others, you now need to set
    the beaming/irradiator parameters to False.
(4) Many speed optimizations: circular orbits (with all effects included) are
    around 7 times slower than WD, well within our release goals. Eccentric orbits
    are still a little over 10 times slower
(5) Many documentation updates and code improvements/cleanups
(6) With all the changes, it is possible that some old cases don't work as
    expected anymore. If this happens to you (or even if it doesn't), please send
    me a minimal script of the features that you would like to have tested in
    the testlib (I always test that library before committing).